requirements.txt
pandas
Pillow
requests
python-dotenv
openpyxl